Moving *all* integration code into SimpleIntegrator.
Renamed BasicIntegrator -> SimpleIntegrator, and pulled out
all the code that's being used for updating the position,
velocity, or acceleration of any body into that class.  This will
give us enough of a basis to start writing new integrators that
don't depeond on any specific code in the body objects themselves.