Added the Verlet integration.
Added a Verlet integrator in the class jctd.physics.VerletIntegrator,
and modified Simulation so that it uses it by default.  Updated Body,
LocatedBody, and DynamicBody to store the previous Location value and the
veloctiy and acceleration vectors, as a way to support the Verlet integrator.
Updated Location to include methods for initializing from, and converting to,
Vector objects.

The integrator doesn't seem to be doing what I want yet, though.